alemdjango/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
atomicwrites==1.4.0
attrs==19.3.0
Babel==2.9.0
beautifulsoup4==4.8.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.4
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.2
django-admin-sortable2==0.7.6
django-allauth==0.40.0
django-angular==2.3
django-appconf==1.0.3
django-classy-tags==1.0.0
django-cms==3.7.2
django-currentuser==0.5.2
django-email-auth==0.1.5
django-entangled==0.3
django-filer==1.7.1
django-filter==2.4.0
django-formtools==2.2
django-fsm==2.7.0
django-fsm-admin==1.2.4
django-ipware==3.0.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-parler==2.0.1
django-parler-rest==2.1
django-phonenumber-field==4.0.0
django-polymorphic==2.1.2
django-post-office==3.4.0
django-rest-auth==0.9.5
django-sass-processor==0.8
django-sekizai==1.1.0
django-select2==7.2.3
django-treebeard==4.3.1
djangocms-admin-style==1.3.0
djangocms-bootstrap==1.1.2
djangocms-cascade==1.3.1
djangocms-text-ckeditor==4.0.0
djangorestframework==3.11.0
djangoshop-stripe==1.2
easy-thumbnails==2.7
factory-boy==2.11.1
Faker==4.0.3
filelock==3.0.12
html5lib==1.1
idna==2.10
importlib-metadata==1.6.0
inflection==0.4.0
jsonfield==3.1.0
libsass==0.20.0
lxml==4.5.0
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.3
pathlib2==2.3.5
phonenumbers==8.12.3
Pillow==8.0.1
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.0
pyparsing==2.4.6
pytest==5.3.1
pytest-django==3.7.0
pytest-factoryboy==2.0.3
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==1.53.0
text-unidecode==1.3
toml==0.10.2
tox==3.7.0
Unidecode==1.1.1
urllib3==1.25.9
virtualenv==20.2.1
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.1.0
